Refactoring

+++ src/main/java/ru/iandreyshev/supermarketSimulator/customer/IBuyer.java
@@ -1,15 +1,16 @@
 package ru.iandreyshev.supermarketSimulator.customer;
 
 import ru.iandreyshev.supermarketSimulator.product.Basket;
+import java.math.BigDecimal;
 
 public interface IBuyer {
     BuyerType getType();
 
     PaymentType getPayType();
 
-    int getMoney();
+    BigDecimal getMoney();
 
-    void addBonus(int bonus);
+    void addBonus(BigDecimal bonus);
 
     Basket getBasket();
 }
